Move envelope schema into index.ts, use AssistantEventSchema for message, remove generic
